+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 *
+* Shader program was not applied to SDL output
 
 ## [0.11.5] - 2025-03-18